Memory leak fixed in strxml.cc . Was causing mdpclient's memory usage to grow indefinitely.
Explanation:
If there are some characters (newline, spaces...) between two XML tokens, 
do_node creates a first string in the "top" field of variable ps, which
is likely to be replaced by another string later on. The first string is
not deleted, but lost.

Solution:
If a string replaces the "top" string, delete the previous one.